+++ src/server/consts.ts
@@ -4,12 +4,3 @@ export const RDS_OPEN_IN_EDITOR = `${RDS_PREFIX}/open-in-editor`;
 export const RDS_DEVTOOLS_UPDATE = `${RDS_PREFIX}/devtools-update`;
 export const DEPENDENCY_PREFIX = "@dep";
 export const ENTRY_POINT = "src/index.tsx";
-
-// https://github.com/vitejs/vite/blob/main/packages/vite/src/node/constants.ts#L14-L23
-export const ESBUILD_MODULES_TARGET = [
-  "es2020",
-  "edge88",
-  "firefox78",
-  "chrome87",
-  "safari13",
-];